chore(konnect): segregate entities envtests file by type
